ICC build fixes warning fixes get rid of SAFE_WCSCPY and SAFE_STRCPY
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7239 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/file/io/write_buffer.cpp
@@ -34,7 +34,7 @@ void WriteBuffer::Append(const void* data, size_t size)
 {
 	if(m_size + size > m_capacity)
 	{
-		m_capacity = round_up_to_pow2((size_t)(m_size + size));
+		m_capacity = round_up_to_pow2(m_size + size);
 		shared_ptr<u8> newData = io_Allocate(m_capacity);
 		cpu_memcpy(newData.get(), m_data.get(), m_size);
 		m_data = newData;
